+++ Rogue.Types/Point.cs
@@ -7,14 +7,14 @@
 
         }
 
-        public Point(int x, int y)
+        public Point(float x, float y)
         {
             this.X = x;
             this.Y = y;
         }
 
-        public int X { get; set; }
+        public float X { get; set; }
 
-        public int Y { get; set; }
+        public float Y { get; set; }
     }
 }
